fix: fixed a bug where .gitkeep was being pushed to bucket

+++ utils/save_files.py
@@ -16,18 +16,19 @@ def save_files_to_bucket():
       else:
          for file in os.listdir("./temp"):
 
-            file_path = os.path.join("./temp", file)
-            print(f"Uploading {file_path}")
-            
-            with open(file_path, "rb") as f:
-               response = supabase.storage.from_(bucket).upload(
-                     file,  
-                     f,
-                     {"upsert": "true"}
-               )
-               print(response)
-            
-            os.remove(file_path)
+            if (file.endswith('pdf')):
+               file_path = os.path.join("./temp", file)
+               print(f"Uploading {file_path}")
+               
+               with open(file_path, "rb") as f:
+                  response = supabase.storage.from_(bucket).upload(
+                        file,  
+                        f,
+                        {"upsert": "true"}
+                  )
+                  print(response)
+               
+               os.remove(file_path)
 
    except FileNotFoundError as e:
       print("File not found!\n")
